==> 04_assignmentfile_saved.Rmd <==
---
title: "GW Assignment - Biden Transition"
author: Mollie Wormser
date: December 6, 2020
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)

library(tidyverse)
library(lubridate)
library(janitor)
library(scales)
library(gt)
library(kableExtra)
options(dplyr.summarise.inform = FALSE)

#run the script for step 03 to update the datasets to capture the current live verions
#and load all the new dataset objects into the environment (you'll see them populate on the upper right window)
source("03_compare_current_with_previous.R")


```

**Note: use either kable/kableExtra or gt tables to format your final tables for the questions below. You'll likely find it easier to figure out the data parts first, then once that's done add the table formatting.**

# AGENCY TEAMS

Come up with the necessary R code to return the following for the agency review teams.

### *New Names*

Below write code to show the new names added to the agency review team lists since the prior data provided.  
  

```{r, echo=FALSE}

newnames %>% 
  gt() %>%
  tab_header(
    title = "New Agency Names Review Team"
  )%>%
  cols_label(name="Name", agency = "agency", on_multiple_teams = "on multiple teams", most_recent_employment= "most recent employment", source_of_funding = "source of funding") %>%
  cols_hide(columns = vars(team_lead, idstring, namestring))



```


### *Aggregate Team Changes*

Add data to show the total number of people appointed to each agency team, along with change since last time reflecting the number of new people added. Omit agencies with no change at all.  
  

```{r, echo=FALSE}

agencycount_compare%>%
  filter(change > 0) %>%
  gt() %>%
  tab_header(title = "total people appointed in each agency and change from previous count") %>%
  cols_label(agency = "agency", current_count = "current count", previous_count = "previous count", change = "change")



```


### *Largest and Smallest Overall*

Show the top 10 **largest** agency review teams as of today:


```{r, echo=FALSE}

agencycount_current %>%
  arrange(desc(current_count)) %>%
  filter(current_count > 22) %>%
  gt() %>%
  tab_header(title = "10 largest agency review teams") %>%
  cols_label(agency = "agency", current_count = "current count")



```

Show the top **smallest** agency review teams as of today - which we'll define here as less than five members:

```{r, echo=FALSE}

agencycount_current %>%
  filter(current_count < 5) %>%
  gt() %>%
  tab_header(title = "smallest agency review teams") %>%
  cols_label(agency = "agency", current_count = "current count")

```



# WHITE HOUSE SENIOR STAFF

Come up with the necessary R code to return the following for the WH senior staff.

### *New Names*

Below write code to show the new names added to the senior staff lists since the prior data provided.  


```{r, echo=FALSE}

source("01_scrape_agencyteams.R")
source("02_scrape_nominees.R")

newstaffnames %>%
  gt() %>%
  tab_header (title = "New Seinor Staff Names") %>%
  cols_label(name = "name", title = "title") %>%
  cols_hide(columns = vars (link, idstring))

```


### *Total Number of Current vs. Previous*

Add code to show the total number of people currently named to the WH senior staff, vs. the previous total number.  
  

```{r, echo=FALSE}
nrow(staff_data_current)
nrow(staff_data_previous)


staff_data_current

staffcount_current <- staff_data_current %>%
  tally(n= "number of people in WH Senior Staff")

staffcount_current

staffcount_previous <- staff_data_previous %>%
  tally (n= "number of people previously WH senior staff")


staffcount_previous

staffcountcompare <- merge(staffcount_current, staffcount_previous, all= TRUE)
staffcountcompare

staffcountcompare %>%
  gt() %>%
  tab_header(
    title = "total current WH senior staff vs. previous WH Senior Staff"

  )



```
